Fix multi-crop set wallpaper dialog rotation issue
When rotating in set wallpaper dialog, the activity restarts, and the
new previews are binded incorrectly, causing the small preview to no
longer be clickable. This is partially due to the discrepancy in
nav destination state. When the view is destroyed due to activity
restart, onDismiss is called on set wallpaper dialog, which pops back
stack, but there is a latency. Therefore, after restart, the initial
state from nav controller still indicates set wallpaper dialog is on the
screen. In addition, we bind all previews the same way, according to the
nav destination state, therefore all previews, in small preview fragment
and set wallpaper dialog, were binded as if they were in set wallpaper
dialog. Instead of relying on the nav dest from the nav controller,
differentiate based on the calling view binder.

Flag: com.android.window.flags.multi_crop
Bug: 369976049
Test: manually verified with new picker UI flag on and off
Test: manually verified on large screen and foldable devices
Change-Id: I936aa200bb7716dbff559c38dc01bddd101aae41